Make installer failure tests snapshot deterministic
