+++ Tests/KeyPathTests/InstallationEngine/InstallerEngineEndToEndTests.swift
@@ -527,41 +527,3 @@ private final class StubHelperMaintenance: WizardHelperMaintaining {
         return true
     }
 }
-
-@MainActor
-private final class StubSystemValidator: WizardSystemValidating {
-    private let snapshots: [SystemSnapshot]
-    private var nextSnapshotIndex = 0
-    private(set) var freshnessRequests: [WizardSystemSnapshotFreshness] = []
-    private(set) var cacheInvalidationCount = 0
-
-    init(snapshot: SystemSnapshot) {
-        snapshots = [snapshot]
-    }
-
-    init(snapshots: [SystemSnapshot]) {
-        precondition(!snapshots.isEmpty)
-        self.snapshots = snapshots
-    }
-
-    func checkSystem() async -> SystemSnapshot {
-        nextSnapshot()
-    }
-
-    func checkSystem(freshness: WizardSystemSnapshotFreshness) async -> SystemSnapshot {
-        freshnessRequests.append(freshness)
-        return nextSnapshot()
-    }
-
-    func invalidateCaches() {
-        cacheInvalidationCount += 1
-    }
-
-    private func nextSnapshot() -> SystemSnapshot {
-        let snapshot = snapshots[nextSnapshotIndex]
-        if nextSnapshotIndex < snapshots.count - 1 {
-            nextSnapshotIndex += 1
-        }
-        return snapshot
-    }
-}
